refactor: move to helpers namespace
